refactor: AlertHistoryRepository 위치 이동
- 다른 도메인들 처럼 Repository가 /domain에 있도록 수정 (일관성 측면)
